telnetd: Try harder to initialize curses
We actually don't care about setting up curses -- we're not going to use
it. We just want it to initialize its datastructures and stuff.  I don't
know of a way to do this besides calling setupterm(), though.
Unfortunately, this can fail when sys.stdout isn't as expected.  We now
try a bit harder by falling back to sys.__stdout__.  This isn't a great
solution.