gephi_topo: Actually use HTTP; refactor
Previously, we just threw the graph data directly at the socket even
though it was supposed to be in an HTTP session.  This seemed to work fine
in all my initial tests, but it has seemed to cause trouble for at least
one or two other people (probably due to later Java versions).

This commit actually adds some rudimentary HTTP handling, which seems to
solve the problem.  It's possible we should switch to using POX's web
infrastructure in the future, but this hasn't been explored yet.

This commit also refactors the code slightly to put more details of the
communication in the hands of the worker class.

Thanks to Felician Nemeth for his assistance in tracking down the issue
and constructing and testing the fix.