of_01: Allow launching of_01 multiple times
This commit allows you to launch of_01 multiple times so that you can
listen on multiple addresses.  The instances register on core as "of_01",
followed by "of_01-2", "of_01-3", etc.  This may not be the best
solution; suggestions/patches welcome.

It also required a slight change to boot.py as part of a minor refactor
of the demand-loading code for of_01 (see 4389e53aa).