of_01: Don't abort OpenFlow loop when accept fails
The only functional improvement here is that the OepnFloow loop no
longer aborts when accept() fails because we've run out of file
descriptors.  However, the error handling code was also restructured in
such a way as to hopefully make it easier to add more special-case
error handling in the future.