openflow.nicira: Use custom OF message handlers
The new OpenFlow message handler mechanism both broke the fairly
hacky way the nicira module used to hook vendor messages and also
allows for a more elegant solution.  This commit rectifies things.

Specifically, it creates a new handler subclass for switches with
Nicira extensions, which it enables in a ConnectionHandshakeComplete
handler.

Signed-off-by: Murphy McCauley <murphy.mccauley@gmail.com>