openflow: Request switch desc on connection
It can be useful to know early on what type of switch you're connected to.
For example, you may need/wish to do special initialization if a switch
supports some extension.  To facilitate this, we now send a switch desc
stats request as part of handshaking by default (it can be disabled
relatively easily), and the result shows up as Connection.description which
should be available by ConnectionUp time.