Add commandline option for epoll
As far as I know, nobody currently cares about this.  Nevertheless, this
was brought up in a comment on #27, and I thought I might as well
implement the feature before closing the issue.

Basically, on systems where epoll is supported, you can use --epoll-sh to
make recoco use epoll instead of select under the hood.  It's probably
*less* efficient due to maintaining select semantics, but it does allow
you to use huge numbers of file descriptors.