edge.aggregator: Fix RX_REMOTE_TABLE
When a source switch has a packet to send to a destination switch, it
tags the packet with the destination port number.  When a destination
switch gets such a packet, it reads the destination port out of the tag
and sends the packet out the appropriate local port.

The question is: which port numbers are in the tag?  The global port
numbers of the aggregate switch, or the local port numbers of the
destination switch?  Either one is feasible.  Prior to this commit, though,
the source and destination disagreed -- the source was putting the local
port number of the destination in the packet, but the destination was
expecting the global port number.  Unsurprisingly, this didn't work well.
Now, the port number in the packet is the local port number of the
destination.  This is chosen because bits in the tag may be at a premium.
Using the local port number, a smaller number of bits in the tag only
limits the number of ports on a single switch instead of the total number
of ports (globally) that the system supports.