of_01: More robust handler-setting
While the previous commit fixes a specific bug of a mismatched handler
name, this commit attempts to make the matching of handler to OpenFlow
message more robust to such errors in general.  It uses the handler name
to find the OpenFlow message type rather than the reverse.  (This is the
technique I proposed on pox-dev in response to the initial bug report.)

Thanks again to Peter Peresini for reporting the initial bug.