Defer delivering port_status received during handshake
POX currently drops port_status messages during the switch--controller handshake.
Unfortunately this leads to a race condition where important topology changes
can occur between the time that the features_reply message is received--
containing a complete list of ports at the moment it was sent--and the time
the handshake completes and port_status messages begin to be delivered.

This patch collects port_status messages received on a connection prior to the
switch--controller handshake completing and defers processing them, and sending
port_status events, until the connection completes.