Defer delivering port_status received during handshake
The set of OpenFlow message handlers active during handshaking
silently ignores port_status messages.  If port_status messages come
before the features reply, this should be okay -- the ports in the
features reply should be fine.  But if they come afterwards, we've
potentially lost information about ports.  Therefore, this commit
buffers port_status messages between the features reply and the
completion of the handshake, and processes them as the hanshake
completes.

It might actually make sense to update the ports in the connection's
ports collection *before* raising the ConnectionUp event and simply
ignore the port state events, but we don't currently do this.

Addresses issue #162.  Based on a patch by @samueljero  (#163).