revent: Add helpful comment
If you try to add a plain function as a weak listener, it fails, which
is okay because this doesn't make any sense.  This just adds a comment
to the line on which it fails so that if you're reading the stack
trace, it might serve to make it clearer.