config: Add config file module
After like ten years, POX finally has a supported config file module.
Essentially, config files are the exact equivalent of the commandline.
You specify modules as sections, with arguments as keys or key-value
pairs, in an INI-like file.  Specifying a module will cause it to
be loaded, and keys become arguments.  There are also a couple
special directives like "!ignore" which causes a section to be ignored.

Example config file:
[samples.pretty_log]
DEBUG

[forwarding.l2_learning]

[openflow.of_01]
port=6653