proto: Add linux_rip RIP router component
This is a non-OpenFlow component which implements RIPv4 and configures
the machine's routing tables accordingly.

It could use lots of improvements.

It actually originally stems from an OpenFlow RIP implementation, but
it has been improved pretty substantially.  The RIP protocol parts
should be factored out and used to implement a new OpenFlow version.