boot: Do foo=bar component name handling earlier
In the fairly recent history, support was added for a shortcut, where
you could specify the first argument to a module without needing to
put the parameter's name.  But it did the check too late.  For example,
it did it after a custom launcher name would have been split off of the
component name.  This commit does it earlier.