interfaceio: Refactor tap reads
Rather than have the IO loop do the actual read, it now simply calls the
associated TapInterface and has *it* do its own read.  Additionally, the
size of the read is no longer hardcoded but is a TapInterface property.
In theory, this should probably always be MTU+4 or something like that,
but fetching this involves a syscall, etc., so we don't currently really
do it dynamically.  But at least it's adjustable.