discovery: Discover upped links much faster
We now send a discovery packet whenever a port is modified and isn't
down.  The idea is that if a port had been administratively down, and
then comes up, we immediately try to discover the link.  We also now
put this port at the *start* of the next discovery cycle instead of
the end.