rip: More flexibile packing of messages
Previously, the RIP core's get_responses() was hard coded to put up
to 25 RIP entries in each RIP response message.  First off, we now
put the handling of this into a method which can be overridden.
Secondly, we now actually try to pack it to meet a particular MTU,
which can be set explicitly as a parameter to get_responses().