revent: Handlers default to priority 0 (not None)
Previously, handlers defaulted to priority None, which is less than any
integer, so they'd always be called last.  This meant that it wasn't
possible to set another handler to be called *after* a default one (not
reliably, anyway).  Now, handlers default to priority 0, so it's totally
possible -- just add handler with priority -1.

The cost here is mostly an extra hash lookup at handler-add time.  This
could be swapped for an extra list lookup at add and event-raise time...