datapaths.pcap_switch: Add virtual ports
This adds virtual ports to the PCap switch.  It's sort of like the OVS
patch panel.  If you have multiple switches running in the same POX
instance, you can use virtual ports to wire them together.

When instantiating a switch from the commandline, port names starting
with a "@" are virtual ports (the "@" is stripped from the eventual
name).  A virtual port is connected to a channel.  If you have multiple
virtual ports connected to the same channel, packets are passed
between them.  If the name you pass in when instantiating the switch
has a *second* "@" in it, the part after the second one is a
channel name.  Thus, you could make two switches connected to each
other with a config file like:

  [datapaths.pcap_switch]
  ports=@eth1@A

  [datapaths.pcap_switch]
  ports=@eth1,@eth2@A

That connects the first switch's eth1 to the second switch's eth2 via
a virtual channel called "A".  And the second switch has a virtual
port called eth1 which isn't connected to anything.

Note: These should really be available to other switch subclasses too.