py: Don't start if POX already quit
If POX is shutdown before the Python console has come up, the
console would still come up even though POX was down.  We now
check at the last moment and don't start up the console if
POX is down.  This is racy, but I'm not going to bother fixing
it at the moment since the Python console is already racy in
much worse ways.