py: Add "sync" mode that syncs with scheduler
Since the beginning of time, running code in the Python console in
POX has been racy since the console runs in a separate thread than
the POX scheduler.

There's now a console variable "sync" (initial value set with --sync).
If it's True, the console syncs with the scheduler before running
code, finally making it safe.

It defaults to False, which maintains the previous behavior.  The
concern with enabling it by default is that it may be possible to
lock up the console.