recoco: Scheduler.synchronized() can be nested
Previously, if you used nested Scheduler.synchronized()s, you'd
deadlock because the second one was waiting for the scheduler
to run, but it's frozen by the first one.

Scheduler.synchronize() now uses a single Synchronizer per
thread (saved in thread local storage), and Synchronizers
are smart enough not to deadlock on themselves, so problem
solved.  (I think.)