recoco: Add task_function decorator
This also adds some related stuff, notably the Again BlockingOperation.

A limitation of Tasks for a long time has been that you can't (easily)
have a Task call a function which calls additional BlockingOperations.
This is now possible by decorating the to-be-called method with
@task_function, and calling it using yield like you do with
BlockingOperations.  Inside the method, replace "return" with "yield".