web: Web POX goes down, shut down open sockets
POX wasn't always quitting in a timely manner web the web server had
been in use.  It should now.

Modern browsers (especially Chrome) may open (or leave open) HTTP
connections without sending a request for quite a while.  Python's
webserver will open requests for these which will then just block at the
readline() in handle_one_request().  The downside here is that when POX
tries to shut down, those threads are left hanging.  We could change
things so that it didn't just blindly call and block on readline.  Or we
could make the handler threads daemon threads.  But instead, we just
keep track of the sockets.  When POX wants to go down, we'll shutdown()
the sockets for reading, which will get readline() unstuck and let POX
close cleanly.