web: Improve startup log message
In many terminals now, URLs are clickable, so the log message now
looks like a URL.  This also communicates whether https is enabled
now that we support https.  We also now log it at INFO instead of
DEBUG, which probably makes more sense, and improves the chances
that you can actually see it to click on.