of01: Allow connections from higher OpenFlow versions
Later OpenFlow versions now use a bitmap of supported protocols.  This
changes the size of the hello message which used to cause us to
immediately abort the connection.  We now let the connection proceed.

If the switch supports OpenFlow 1.0, it should downgrade the connection.
We now check that the features reply is actually an OpenFlow 1.0 message
and kill the connection if it isn't.  However, this probably won't get
triggered because the only logical thing for a switch that doesn't
support 1.0 to do in this case is disconnect.  OVS, at least, does this.