web: Add POX CookieGuard CSRF mitigation
noxrepo/pox#240 points out that the OpenFlow webservice is vulnerable
to CSRF attacks.  This commit adds a mitigation feature which we dub
POX CookieGuard.  The POX webserver sets a SameSite=Strict cookie and
demands it be there in future requests.  In the case of a CSRF attack,
the attacker cannot make this happen assuming the attack is launched
from a different origin (if the attacker can make the request happen
from the POX webserver origin, I think you probably have bigger
problems).  Note that this protects *all* requests through the POX
webserver -- not just the OpenFlow webservice.

POX CookieGuard is enabled by default, but can be disabled using
"web --no-cookieguard".