host_tracker: Fix expiration warnings
When an entry expires, it seems like it should have no IP addresses.
There's code that logs a warning for any IP address that still exists,
but it looks like it was broken and just logged the wrong thing.
This same code removes the existing ones, which was also broken.