py3/telnetd: Fix str/bytes
This was a bit more involved than a lot of fixes.  The basic breakdown
is that the telnet worker works in raw bytes, but personalities work
in strings.  The personalities now have a little bit of encode/decode
logic (and CRLF translation), as well as a new handler for when
decoding off the wire results in a bad value.