web: Make websockets work better with cookieguard
If a websocket tries to connect and cookieguard intercepts it and sends
a 307, at least some browsers (Chrome, for example) won't follow it, and
*won't even accept cookies* from the 307, so the cookieguard cookie
won't get set.

This commit makes it so that even when cookieguard fails, the websocket
connects.  If cookieguard fails, it's then immediately disconnected,
but the cookie at least gets set this way.  Thus, if the websocket
reconnects, it'll then pass cookieguard.  If the application has
retry logic, this will hopefully appear to just work.  If the user can
make it retry, that will work too (with slightly worse UX in the case
where it doesn't work the first time).

This shouldn't actually come up much, since generally a websocket
will be accessed via a page, and the act of loading the page will
pass cookieguard.  But if the websocket is used by a non-webpage,
it can happen.  It can also happen if you restart POX while the
page is open (which may not be uncommon during development).