web: CookieGuard consumes POST before redirecting
If one were to do a file upload with a bad cookie, POX may have
redirected before the file upload was complete, causing a connection
reset which stops the file upload process.  This could probably
happen with a normal form too, at least if it's long enough.

The default is now to actully read all the POST data and *then*
redirect.  This will require the file to be uploaded three times,
but it won't break, which is assumed to be better in general.