recoco: Remove unneeded call to super __del__
Scheduler tried to call its superclass's __del__.  Apparently
this worked or silently failed in Python 2, and maybe in newer
version of Python 3, but in 3.6.9, this causes an exception.
The exception is harmless, but dumb, so this commit just
removes the call.  Scheuler's superclass is "object", so it's
not necessary anyway.