dhcpd: Factor core functionality into superclass
DHCPD used to specifically work with OpenFlow datapaths.  This commit
pulls the basic server into a new DHCPBase class so that you can
easily create subclasses that work with whatever (e.g., that run
on a tap interface).  The OpenFlow-specific version is now a subclass
(but could be broken, since I haven't tested it).