revent: autoBindEvents() is more careful
autoBindEvents used to iterate over all the attributes of the instance
it was given.  This has always had the unfortunate side effect of
iterating over everything defined as a @property too -- and therefore
running the property code, potentially very early on in the object's
lifetime.

Instead, we now iterate over the attributes of the instance's *class*.
For things that look promising as an event handler (they're callable),
we *then* look at the instance.  A property isn't a callable on the
class, so this gets around the aforementioned problem.

An alternate approach would be checking the attribute name to see if
it looks like a handler before checking if it's callable.