core: Improve some going-up logic
Really, the whole going up (and going down) logic could use a complete
rewrite.  It's gotten messy, and the deferral idea is useful enough that
it should probably be available for going down as well as up.

For now, though, this attempts to at least stop a corner case from
causing the up event to fire twice sometimes!