web: Add support for CORS
This adds some basic support for the Access-Control-Allow headers.
In particular, this has some features for "CORS", including a
permissive mode which is useful for debugging.

I developed this a while ago, and I'm not sure this is the final
version.  Unfortunately, the machine it actually ran on is down for
maintenance for maybe another month or so.  I figured it's better
to push this now than wait for weeks (at which point I will
probably have forgotten about it).