Python 2/3 compatiblity fix on dns decoder.
This commit introduces a compatibility fix in the dns packet decoder.
This prevents the output from being spammed by failures to decode DNS
requests (and presumably responses). The fix is tested under Python
3.9.2 on Debian 11, but steps were taken to retain Python 2
compatibility as well.

The original problems stems from how Python 2 and 3 handle `bytes`
objects differently - Python 2 seems to just handle them as `str`
objects, whereas under python 3 they are `bytes`. This obviously leads
to issues where conversion happen - in this case `bytes` to `int` and
`str`. Whereas under Python 2, `ord($OBJ)` is required to convert a
byte to int, under Python 3 it is sufficient to just use the byte as
`int`. Similarly, sliced bytestring under Python 2 is already a `str`
object, whereas in Python 3 it requires `.decode()`ing first (in Python
2 you can decode a `str` object, but it yields a `unicode` object,
which might not be desirable).